+++ test/collateral/html/2d.gradient.html
@@ -36,6 +36,8 @@
 <script type="text/javascript" src="asserts.js"></script>
 
 <script type="text/javascript">
+    window.ImageData = ImageData;
+
     async function getAsset(name) {
       return new Promise((resolve, reject) => {
         let img = document.createElement("IMG")
@@ -84,6 +86,7 @@
 
     function focusTest (num) {
       window.ctx = testContexts[num][0]
+      window.allContexts = testContexts[num]
     }
     window.onhashchange = () => {
       var url = new URL(window.location)
@@ -109,6 +112,7 @@
       var url = new URL(window.location)
       if (url.searchParams.get("runTests") != null) {
         selected_test_ids = url.searchParams.get("runTests").split(",").map((x) => parseInt(x))
+        selected_test_ids = selected_test_ids.filter((elm, idx) => {return isFinite(elm)}).sort()
         tests = selected_test_ids.map((x) => document.getElementById("test_"+x+"_content")) 
       }
 
@@ -1000,7 +1004,6 @@
 
 <hr />
 
-<!-- DISABLED: TODO support cross-canvas drawing
 <p><a id="305" href="#305"\>305</a>: 2d.gradient.object.crosscanvas</p>
 
 <script>
@@ -1038,8 +1041,6 @@
 [<a href="?runTests=305&noCleanup#305" onclick="window.location.reload()">solo run</a>]
 
 <hr />
-
--->
 
 <p><a id="303" href="#303"\>303</a>: 2d.gradient.object.update</p>
