+++ test/collateral/html/2d.missingargs.html
@@ -36,6 +36,8 @@
 <script type="text/javascript" src="asserts.js"></script>
 
 <script type="text/javascript">
+    window.ImageData = ImageData;
+
     async function getAsset(name) {
       return new Promise((resolve, reject) => {
         let img = document.createElement("IMG")
@@ -84,6 +86,7 @@
 
     function focusTest (num) {
       window.ctx = testContexts[num][0]
+      window.allContexts = testContexts[num]
     }
     window.onhashchange = () => {
       var url = new URL(window.location)
@@ -109,6 +112,7 @@
       var url = new URL(window.location)
       if (url.searchParams.get("runTests") != null) {
         selected_test_ids = url.searchParams.get("runTests").split(",").map((x) => parseInt(x))
+        selected_test_ids = selected_test_ids.filter((elm, idx) => {return isFinite(elm)}).sort()
         tests = selected_test_ids.map((x) => document.getElementById("test_"+x+"_content")) 
       }
