+++ utilityObjects.js
@@ -1,0 +1,1 @@
+export var ImageData = window.ImageData;
